third_party/WebKit/LayoutTests/http/tests/security/isolatedWorld/bypass-main-world-csp-font-src-expected.txt

ALERT: Bypass main world's CSP with font-face.
This test ensures that scripts run in isolated worlds marked with their own Content Security Policy aren't affected by the page's font-src directive.
